details about merge and commits
* why
I found myself dealing with those details in my first weeks.
The first one is just to clarify that remove unnecessary commits in a pull request doesn't mean to
keep all the development in one commit.
I think is better to have several commits with atomic changes than one commit with
everything.
The second one is about merging, I consider that is important to know why is
better to do it from the console.

* How
Proposing this PR to debate!